refactor(ui): address code review feedback
- Fix palette selection feedback double-processing (Comment 1):
  - Add base_color_for() helper to centralize material→color resolution
  - Make selection_changed and has_new_previews branches mutually exclusive
  - Avoid processing Added<PalettePreview> twice when selection changes

- Fix ghost preview material fallback (Comment 2):
  - Handle missing material explicitly with warning instead of empty Handle
  - Early return to despawn ghost when no material available

- Move UI asset initialization to UiPlugin (Comment 3):
  - Create setup_ui_assets() startup system in ui/mod.rs
  - Move GhostPreviewMaterial and CheatIndicatorTexture init from lib.rs
  - Remove unused asset_server parameter from setup()

- Remove placeholder tests (Comments 4-6):
  - Delete ui_fallible_systems.rs (documentation-only)
  - Delete ui_palette_change_detection.rs (documentation-only)
  - Delete ui_cheat_indicator_caching.rs (documentation-only)
  - Behavior already validated by real integration tests

- Comment 7 (docs typo): Already correct as 'Bevy 0.17'\''s mandates'

All 7 code review comments addressed.
