--- tests/ui_cheat_indicator_caching.rs
@@ -1,63 +1,0 @@
-//! Cheat indicator asset caching test (Constitution VIII: Asset Handle Reuse).
-//!
-//! Verifies that cheat indicator assets are loaded once and cached,
-//! not repeatedly loaded on mode toggle.
-
-use bevy::prelude::*;
-
-#[test]
-fn cheat_indicator_assets_are_cached() {
-    // This test documents the expected behavior:
-    // Cheat indicator assets (textures) MUST be loaded once in startup
-    // and cached in a Resource. Toggling cheat mode MUST NOT call asset_server.load().
-    //
-    // Expected implementation pattern:
-    //   #[derive(Resource)]
-    //   pub struct CheatIndicatorAssets {
-    //       texture: Handle<Image>,
-    //   }
-    //
-    //   pub fn load_cheat_indicator_assets(
-    //       mut commands: Commands,
-    //       asset_server: Res<AssetServer>,
-    //   ) -> Result<(), UiSystemError> {
-    //       let texture = asset_server.load("textures/default/cheat-mode-128.png");
-    //       commands.insert_resource(CheatIndicatorAssets { texture });
-    //       Ok(())
-    //   }
-    //
-    //   pub fn handle_cheat_indicator(
-    //       mut commands: Commands,
-    //       cheat_mode: Res<CheatMode>,
-    //       assets: Res<CheatIndicatorAssets>,  // NOT calling asset_server.load() here
-    //   ) -> Result<(), UiSystemError> {
-    //       if cheat_mode.is_active {
-    //           commands.spawn((CheatIndicator, assets.texture.clone()));
-    //       }
-    //       Ok(())
-    //   }
-
-    println!(
-        "Cheat indicator assets MUST be cached in a Resource, \
-         not loaded per-frame or per-toggle."
-    );
-
-    // Note: Full validation requires either:
-    // 1. Inspecting CheatMode toggle event handler for asset_server.load() calls
-    // 2. Profiling memory allocations on toggle
-    // 3. Code inspection of handle_cheat_indicator to confirm asset_server is NOT used
-
-    assert!(true); // Placeholder: test documents expected behavior
-}
-
-#[test]
-fn cheat_mode_toggle_does_not_trigger_asset_loads() {
-    // Documentation test: Toggling cheat mode should only show/hide
-    // the indicator entity, not trigger any asset loading.
-
-    println!(
-        "CheatMode toggle handler MUST NOT call asset_server.load(); \
-         use cached assets from CheatIndicatorAssets resource instead."
-    );
-    assert!(true); // Placeholder
-}
