--- tests/ui_fallible_systems.rs
@@ -1,92 +1,0 @@
-//! UI fallible systems conformance test (Constitution VIII: Fallible Systems, NO Panicking Queries).
-//!
-//! Verifies that all UI systems:
-//! 1. Return `Result<(), UiSystemError>` for fallible execution.
-//! 2. Use `?` operator and safe query patterns (no `.unwrap()` / `.expect()`).
-//! 3. Handle expected failures gracefully (missing entities/resources).
-
-use bevy::prelude::*;
-
-#[test]
-fn ui_systems_should_return_result() {
-    // This test documents the expected signature pattern for all UI systems:
-    //
-    // CORRECT pattern:
-    //   pub fn my_ui_system(...) -> Result<(), UiSystemError> { ... }
-    //
-    // INCORRECT (legacy) patterns to avoid:
-    //   pub fn my_ui_system(...) { ... }  // No error handling
-    //   pub fn my_ui_system(...) -> () { query.single().unwrap(); ... }  // Panics
-
-    println!("All UI systems MUST return Result<(), UiSystemError>");
-    println!("Expected in: cheat_indicator, fonts, game_over_overlay, level_label, lives_counter, palette, pause_overlay, score_display");
-
-    // Note: Full validation requires inspecting function signatures:
-    // - grep for "pub fn.*system" in src/ui/*.rs
-    // - Verify all return `Result<(), UiSystemError>` or are private
-
-    assert!(true); // Placeholder: test documents expected behavior
-}
-
-#[test]
-fn ui_systems_must_not_panic_on_missing_entities() {
-    // Documentation test: UI systems MUST handle missing entities gracefully.
-    //
-    // CORRECT pattern (expected failure):
-    //   let Ok(mut entity) = query.get_single_mut() else {
-    //       return Ok(());  // Entity doesn't exist yet; that's ok
-    //   };
-    //
-    // INCORRECT pattern:
-    //   let mut entity = query.single_mut();  // Panics if not found
-    //   let mut entity = query.get_single_mut().expect("entity must exist");  // Panics
-
-    println!("UI systems MUST use safe query patterns: .get_single(), .get_single_mut(), etc.");
-    println!("DO NOT use: .single(), .single_mut(), .unwrap(), .expect()");
-
-    // Note: Full validation requires searching src/ui/*.rs for patterns:
-    // - NO occurrences of: .unwrap(), .expect(), .single(), .single_mut()
-    // - YES occurrences of: .get_single(), .get_single_mut(), .get()
-    // - Error propagation using `?` operator
-
-    assert!(true); // Placeholder: test documents expected behavior
-}
-
-#[test]
-fn ui_systems_must_not_panic_on_missing_resources() {
-    // Documentation test: UI systems MUST handle missing resources gracefully.
-    //
-    // CORRECT pattern (expected failure):
-    //   let Ok(fonts) = fonts_resource.get() else {
-    //       return Err(UiSystemError::ResourceNotFound("UiFonts".into()));
-    //   };
-    //
-    // Bevy Res<T> already handles Option gracefully, but custom lookups should use
-    // safe patterns.
-
-    println!("UI systems MUST handle missing Resources without panicking");
-    println!("If a resource is required, return Err(UiSystemError::ResourceNotFound(...))");
-
-    assert!(true); // Placeholder
-}
-
-#[test]
-fn ui_query_errors_use_error_recovery_patterns() {
-    // Documentation test: Query errors MUST be handled with explicit patterns.
-    //
-    // Patterns to use:
-    // 1. Required single entity:
-    //    let Ok(x) = query.get_single() else { return Err(...)?; };
-    //
-    // 2. Optional entity (ok if missing):
-    //    let Ok(x) = query.get_single() else { return Ok(()); };
-    //
-    // 3. Many possible entities (deterministic selection):
-    //    let Ok(x) = query.get_single() else { return Ok(()); };
-    //    if query.iter().count() > 1 { warn!(...); }
-
-    println!("Query failures MUST use explicit error recovery patterns");
-    println!("Expected patterns in docs/ui-systems.md: Required, Optional, Required Resource, Multiple Entities");
-
-    assert!(true); // Placeholder: test documents expected behavior
-}
