--- tests/ui_palette_change_detection.rs
@@ -1,64 +1,0 @@
-//! Palette change-detection test (Constitution VIII: Change Detection).
-//!
-//! Verifies that palette selection feedback is driven by data changes,
-//! not executed every frame when selection hasn't changed.
-
-use bevy::prelude::*;
-
-#[test]
-fn palette_selection_feedback_is_change_driven() {
-    // This test documents the expected behavior:
-    // When SelectedBrick does not change, update_palette_selection_feedback
-    // MUST NOT mutate any UI components (no per-frame work).
-    //
-    // This is a documentation test for now, showing the expected pattern.
-    // Full behavior validation requires running the game and inspecting logs/performance,
-    // or implementing a test system that tracks mutations.
-    //
-    // Expected implementation pattern:
-    //   pub fn update_palette_selection_feedback(
-    //       mut query: Query<&mut BackgroundColor, With<PalettePreview>>,
-    //       selected: Query<&SelectedBrick, Changed<SelectedBrick>>,
-    //   ) -> Result<(), UiSystemError> {
-    //       let Ok(_) = selected.get_single() else {
-    //           return Ok(());  // No change => no work
-    //       };
-    //       for mut color in &mut query { ... }
-    //       Ok(())
-    //   }
-
-    println!(
-        "Palette selection feedback MUST be driven by Changed<SelectedBrick>, \
-         not executed every frame."
-    );
-
-    // Note: Full validation requires either:
-    // 1. Stateful test system that counts mutations between Changed<SelectedBrick> cycles
-    // 2. Performance profiling (ensure palette updates < 0.1ms per frame when selection stable)
-    // 3. Code inspection of `update_palette_selection_feedback` signature
-
-    assert!(true); // Placeholder: test documents expected behavior
-}
-
-#[test]
-fn palette_preview_spawning_uses_added_filter() {
-    // Documentation test: Palette previews should only spawn once when
-    // PaletteRoot is added, not re-spawn every frame.
-    //
-    // Expected implementation pattern:
-    //   pub fn ensure_palette_ui(
-    //       mut commands: Commands,
-    //       palette_state: Res<PaletteState>,
-    //       added: Query<&PaletteRoot, Added<PaletteRoot>>,
-    //   ) -> Result<(), UiSystemError> {
-    //       if !palette_state.open { return Ok(()); }
-    //       let Ok(_) = added.get_single() else {
-    //           return Ok(());  // Already spawned
-    //       };
-    //       // spawn previews
-    //       Ok(())
-    //   }
-
-    println!("Palette UI spawning MUST use Added<PaletteRoot> filter to avoid re-spawning.");
-    assert!(true); // Placeholder
-}
